requirements.txt
aiohttp==3.7.3
asana==0.10.3
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
idna==2.10
idna-ssl==1.1.0
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
lxml==4.6.2
mailchimp-marketing==3.0.28
MarkupSafe==1.1.1
multidict==5.1.0
oauthlib==3.1.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pyee==7.0.4
Pygments==2.8.0
python-dateutil==2.8.1
python-dotenv==0.15.0
repackage==0.7.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
slackclient==2.9.3
slackeventsapi==2.2.1
soupsieve==2.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
yarl==1.6.3
